agentic-engineering-team/requirements.txt
crewai>=0.28.0
docker>=6.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-html>=3.0.0
requests>=2.25.0
flask>=2.0.0

# Development & Linting
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0
pylint>=2.17.0
pre-commit>=3.0.0